pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minilog"
version = "0.1.11"
description = "A mini-Prolog logic-programming engine with Ukrainian-language syntax"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "trafilatura>=2.0",
    "pymupdf4llm>=1.0",
    "python-docx>=1.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12",
    "langdetect>=1.0",
]

[project.scripts]
minilog = "minilog.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["minilog*"]
